:recycle:refactor: Mudança na arquitetura adicionando o módulo de DTO exclusivo
