+++ votify-api/src/main/java/br/com/votify/api/controller/users/UserController.java
@@ -1,14 +1,14 @@
 package br.com.votify.api.controller.users;
 
 import br.com.votify.api.configuration.SecurityConfig;
-import br.com.votify.api.dto.ApiResponse;
-import br.com.votify.api.dto.users.UserDetailedViewDTO;
-import br.com.votify.api.dto.users.UserLoginDTO;
-import br.com.votify.api.dto.users.UserRegisterDTO;
 import br.com.votify.core.domain.entities.tokens.AuthTokens;
 import br.com.votify.core.domain.entities.users.User;
 import br.com.votify.core.utils.exceptions.VotifyException;
 import br.com.votify.core.service.UserService;
+import br.com.votify.dto.ApiResponse;
+import br.com.votify.dto.users.UserDetailedViewDTO;
+import br.com.votify.dto.users.UserLoginDTO;
+import br.com.votify.dto.users.UserRegisterDTO;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
