+++ votify-dto/src/main/java/br/com/votify/dto/users/UserRegisterDTO.java
@@ -1,6 +1,6 @@
-package br.com.votify.api.dto.users;
+package br.com.votify.dto.users;
 
-import br.com.votify.api.dto.DTOInput;
+import br.com.votify.dto.DTOInput;
 import br.com.votify.core.domain.entities.users.CommonUser;
 import br.com.votify.core.domain.entities.users.User;
 import lombok.AllArgsConstructor;
